# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 6016ab4ba3e049112323e4ce
  title: CMO
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        "200":
          description: OK
  /account:
    post:
      x-skaffolder-id: 6016b801a3e049112323e5b9
      x-skaffolder-name: create
      x-skaffolder-id-resource: 6016b20da3e049112323e57e
      x-skaffolder-resource: Account
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Account
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Description:
                  type: string
                MaxCredit:
                  type: integer
                Name:
                  type: string
    get:
      x-skaffolder-id: 6016b801a3e049112323e5bf
      x-skaffolder-name: list
      x-skaffolder-id-resource: 6016b20da3e049112323e57e
      x-skaffolder-resource: Account
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Account
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
  "/account/{id}":
    delete:
      x-skaffolder-id: 6016b801a3e049112323e5be
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 6016b20da3e049112323e57e
      x-skaffolder-resource: Account
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Account
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 6016b801a3e049112323e5ba
      x-skaffolder-name: get
      x-skaffolder-id-resource: 6016b20da3e049112323e57e
      x-skaffolder-resource: Account
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Account
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    post:
      x-skaffolder-id: 6016b801a3e049112323e5b8
      x-skaffolder-name: update
      x-skaffolder-id-resource: 6016b20da3e049112323e57e
      x-skaffolder-resource: Account
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Account
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Account"
  "/account/findByLessor/{key}":
    get:
      x-skaffolder-id: 6016b7dba3e049112323e5b0
      x-skaffolder-name: findByLessor
      x-skaffolder-id-resource: 6016b20da3e049112323e57e
      x-skaffolder-resource: Account
      x-skaffolder-crudAction: findByLessor
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findByLessor
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findByLessor/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findByLessor
      tags:
        - Account
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
...
  /user:
    post:
      x-skaffolder-id: 6016ab4ba3e049112323e4dd
      x-skaffolder-name: create
      x-skaffolder-id-resource: 6016ab4ba3e049112323e4d8
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 6016ab4ba3e049112323e4e0
      x-skaffolder-name: list
      x-skaffolder-id-resource: 6016ab4ba3e049112323e4d8
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 6016ab4ba3e049112323e4dc
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 6016ab4ba3e049112323e4d8
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 6016ab4ba3e049112323e4de
      x-skaffolder-name: get
      x-skaffolder-id-resource: 6016ab4ba3e049112323e4d8
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 6016ab4ba3e049112323e4df
      x-skaffolder-name: update
      x-skaffolder-id-resource: 6016ab4ba3e049112323e4d8
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Account:
      x-skaffolder-id: 6016b20da3e049112323e57e
      x-skaffolder-id-db: 6016ab4ba3e049112323e4cf
      x-skaffolder-id-entity: 6016b20da3e049112323e57a
      x-skaffolder-url: /account
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Description:
          type: string
          x-skaffolder-id-attr: 6016b20da3e049112323e57c
          x-skaffolder-type: String
          x-skaffolder-required: true
        MaxCredit:
          type: integer
          x-skaffolder-id-attr: 6016b655a3e049112323e59b
          x-skaffolder-type: Integer
          x-skaffolder-required: true
        Name:
          type: string
          x-skaffolder-id-attr: 6016b20da3e049112323e57b
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations:
        Lessor:
          x-skaffolder-id: 6016b20da3e049112323e57d
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 6016b20da3e049112323e57a
          x-skaffolder-ent2: 6016b117a3e049112323e572
      required:
        - _id
        - Description
        - MaxCredit
        - Name
...
    User:
      x-skaffolder-id: 6016ab4ba3e049112323e4d8
      x-skaffolder-id-db: 6016ab4ba3e049112323e4cf
      x-skaffolder-id-entity: 6016ab4ba3e049112323e4d0
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 6016ab4ba3e049112323e4d3
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 6016ab4ba3e049112323e4d4
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 6016ab4ba3e049112323e4d2
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 6016ab4ba3e049112323e4d6
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 6016ab4ba3e049112323e4d5
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 6016ab4ba3e049112323e4d1
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
...